+++ Gemfile
@@ -75,3 +75,4 @@ gem "dotenv-rails", "~> 3.1"
 gem "table_print", "~> 1.5"
 
 gem "stripe", "~> 10.11"
+
