+++ app/controllers/static_pages_controller.rb
@@ -3,9 +3,12 @@ class StaticPagesController < ApplicationController
   end
 
   def send_contact
-    UserMailer.visitor_contact_email(params).deliver_now
-    UserMailer.admin_contact_email(params).deliver_now
-    flash[:success] = "Votre message a bien été envoyé."
+    UserMailer.visitor_contact_email(params[:static_pages]).deliver_now
+    if UserMailer.admin_contact_email(params[:static_pages]).deliver_now
+      flash[:success] = "Votre message a bien été envoyé."
+    else
+      flash[:alert] = "Erreur lors de l'envoi. Veuillez recommencer."
+    end
     redirect_to :contact
   end
