+++ app/mailers/user_mailer.rb
@@ -5,8 +5,6 @@ class UserMailer < ApplicationMailer
     #je récupère l'instance user pour ensuite pouvoir la passer à la view en @user
     @user = user 
 
-    #je définis une variable @url qu'on utilisera dans la view d’e-mail
-    ########### à définir avant de basculer sur Heroku #####################################
     @url = application_url
     @signin_url = application_url + "/users/sign_in"
 
@@ -20,6 +18,22 @@ class UserMailer < ApplicationMailer
     ############ à définir une fois le paiement stripe fait (.order) #######################
     @order = @user.order.last if @user.order.any?
     mail(to: @admin.email, subject: 'Une nouvelle commande vient d\'être effectuée !')
+  end
+
+  def visitor_contact_email(params)
+    @name = params[:name]
+    @email = params[:email]
+    @message = params[:message]
+    @url = application_url
+    mail(to: "annie.herieau@gmail.com", subject: 'Catysfaction: Nous avons reçu votre message')
+  end
+
+  def admin_contact_email(params)
+    @name = params[:name]
+    @email = params[:email]
+    @message = params[:message]
+    @url = application_url
+    mail(to: "annie.herieau@gmail.com", subject: 'Catysfaction: Nouveau contact')
   end
 end
